Add unimplemented BDD spec
This is the spec as supplied for the test, but with it made into a valid
Gherkin feature file. Currently all steps skipped due to lack of step
definitions.

FluentAssertions added because I plan to use them and it's a far nicer
way of expressing the meaning behind test assertions.