Expand scenarios - add undiscounted multibuy
I'd missed the quantity when implementing add and there isn't a test for
that without also discounting, added a test for adding multiple without
ending up with a discount.

Added names that shows which of the scenarios are expected to trigger a
discount for ease of validation (perhaps with a business person).